Restructured the project into frontend and backend
Originally I did not want to involve the complexity of `npm`, as it is a tool students may not be familiar with. (whereas python pip is a part of the university courses)

As the project keeps growing with more third party dependencies I decided that it would be too cluttered to have all the dependencies in the repo itself. As a bonus the repo only contains OUR code now, and not a bunch of third party code.

I have tried to remove most of the complexity through docker and docker-compose, such that it should be relatively straight forward to getting the project to run.